+++ GRDB/Core/Database.swift
@@ -768,17 +768,6 @@ extension Database {
 // =========================================================================
 // MARK: - Database Schema
 
-/// The protocol for schema cache.
-///
-/// This protocol must not contain values that are valid for a single connection
-/// only, because several connections can share the same schema cache.
-protocol DatabaseSchemaCache {
-    mutating func clear()
-    
-    func primaryKey(_ tableName: String) -> PrimaryKey??
-    mutating func set(primaryKey: PrimaryKey?, for tableName: String)
-}
-
 extension Database {
     
     /// Clears the database schema cache.
@@ -901,6 +890,50 @@ extension Database {
         return primaryKey
     }
     
+    /// The indexes on table named `tableName`; returns the empty array if the
+    /// table does not exist.
+    ///
+    /// Note: SQLite defines no index for INTEGER PRIMARY KEY columns: this
+    /// method does not return any index that represents this unique constraint.
+    ///
+    /// If you want to know if a set of columns uniquely identify a row, prefer
+    /// columns(_:uniquelyIdentifyRowsIn:) instead.
+    func indexes(on tableName: String) -> [IndexInfo] {
+        if let indexes = schemaCache.indexes(on: tableName) {
+            return indexes
+        }
+        
+        let indexes = Row.fetch(self, "PRAGMA index_list(\(tableName.quotedDatabaseIdentifier))").map { row -> IndexInfo in
+            let indexName: String = row.value(atIndex: 1)
+            let unique: Bool = row.value(atIndex: 2)
+            let columns = Row.fetch(self, "PRAGMA index_info(\(indexName.quotedDatabaseIdentifier))")
+                .map { ($0.value(atIndex: 0) as Int, $0.value(atIndex: 2) as String) }
+                .sorted { $0.0 < $1.0 }
+                .map { $0.1 }
+            return IndexInfo(name: indexName, columns: columns, unique: unique)
+        }
+        
+        schemaCache.set(indexes: indexes, forTableName: tableName)
+        return indexes
+    }
+    
+    /// True if a set of columns uniquely identify a row, that is to say if
+    /// there is a unique index on those columns, or if the column is the
+    /// INTEGER PRIMARY KEY.
+    func columns<T: Sequence where T.Iterator.Element == String>(_ columns: T, uniquelyIdentifyRowsIn tableName: String) throws -> Bool {
+        let primaryKey = try self.primaryKey(tableName) // first, so that we fail early and consistently should the table not exist
+        let columns = Set(columns)
+        if indexes(on: tableName).contains({ index in index.isUnique && Set(index.columns) == columns }) {
+            return true
+        }
+        if columns.count == 1,
+            let rowIDColumnName = primaryKey?.rowIDColumn
+            where rowIDColumnName == columns.first! {
+            return true
+        }
+        return false
+    }
+    
     // CREATE TABLE persons (
     //   id INTEGER PRIMARY KEY,
     //   firstName TEXT,
@@ -925,6 +958,18 @@ extension Database {
             notNull = row.value(named: "notnull")
             defaultDatabaseValue = row.databaseValue(named: "dflt_value")!
             primaryKeyIndex = row.value(named: "pk")
+        }
+    }
+    
+    struct IndexInfo {
+        let name: String
+        let columns: [String]
+        let isUnique: Bool
+        
+        init(name: String, columns: [String], unique: Bool) {
+            self.name = name
+            self.columns = columns
+            self.isUnique = unique
         }
     }
 }
@@ -1038,7 +1083,7 @@ final class StatementCompilationObserver {
         let observerPointer = unsafeBitCast(self, to: UnsafeMutablePointer<Void>.self)
         sqlite3_set_authorizer(database.sqliteConnection, { (observerPointer, actionCode, CString1, CString2, CString3, CString4) -> Int32 in
             switch actionCode {
-            case SQLITE_DROP_TABLE, SQLITE_DROP_TEMP_TABLE, SQLITE_DROP_TEMP_VIEW, SQLITE_DROP_VIEW, SQLITE_DETACH, SQLITE_ALTER_TABLE, SQLITE_DROP_VTABLE:
+            case SQLITE_DROP_TABLE, SQLITE_DROP_TEMP_TABLE, SQLITE_DROP_TEMP_VIEW, SQLITE_DROP_VIEW, SQLITE_DETACH, SQLITE_ALTER_TABLE, SQLITE_DROP_VTABLE, SQLITE_CREATE_INDEX, SQLITE_CREATE_TEMP_INDEX, SQLITE_DROP_INDEX, SQLITE_DROP_TEMP_INDEX:
                 let observer = unsafeBitCast(observerPointer, to: StatementCompilationObserver.self)
                 observer.invalidatesDatabaseSchemaCache = true
             case SQLITE_READ: