+++ GRDB/Core/DatabaseSchemaCache.swift
@@ -1,20 +1,43 @@
+/// The protocol for schema cache.
+///
+/// This protocol must not contain values that are valid for a single connection
+/// only, because several connections can share the same schema cache.
+///
+/// Statements can't be cached here, for example.
+protocol DatabaseSchemaCache {
+    mutating func clear()
+    
+    func primaryKey(_ tableName: String) -> PrimaryKey??
+    mutating func set(primaryKey: PrimaryKey?, for tableName: String)
+
+    func indexes(on tableName: String) -> [Database.IndexInfo]?
+    mutating func set(indexes: [Database.IndexInfo], forTableName tableName: String)
+}
+
 /// A thread-unsafe database schema cache
 final class SimpleDatabaseSchemaCache: DatabaseSchemaCache {
     private var primaryKeys: [String: PrimaryKey?] = [:]
+    private var indexes: [String: [Database.IndexInfo]] = [:]
     
     func clear() {
         primaryKeys = [:]
+        indexes = [:]
     }
     
     func primaryKey(_ tableName: String) -> PrimaryKey?? {
-        guard let pk = primaryKeys[tableName] else {
-            return nil
-        }
-        return pk
+        return primaryKeys[tableName]
     }
     
     func set(primaryKey: PrimaryKey?, for tableName: String) {
         primaryKeys[tableName] = primaryKey
+    }
+    
+    func indexes(on tableName: String) -> [Database.IndexInfo]? {
+        return indexes[tableName]
+    }
+    
+    func set(indexes: [Database.IndexInfo], forTableName tableName: String) {
+        self.indexes[tableName] = indexes
     }
 }
 
@@ -32,5 +55,13 @@ final class SharedDatabaseSchemaCache: DatabaseSchemaCache {
     
     func set(primaryKey: PrimaryKey?, for tableName: String) {
         cache.write { $0.set(primaryKey: primaryKey, for: tableName) }
+    }
+    
+    func indexes(on tableName: String) -> [Database.IndexInfo]? {
+        return cache.read { $0.indexes(on: tableName) }
+    }
+    
+    func set(indexes: [Database.IndexInfo], forTableName tableName: String) {
+        cache.write { $0.set(indexes: indexes, forTableName: tableName) }
     }
 }